Replace inline style with `hidden` attribute.
The hidden attribute serves the purpose of hiding content from the
document. We can have multiple display options rather that `block`
specially with responsive design where having different display value
based on the viewport might be required. Switching to the hidden
attribute allow us to control the diplay with ease and shift the display
to the CSS files.

I had to set the display of the hidden element to none with important to
ensure having a spcificity higher than the target elements if there are
other display values applied to them.